fix: Moved 'RedeemConfirm' component inside of 'StakePositionsTable' so that it uses the sorted data as reference for displaying positions to be redeemed
